tasks: Add label coloring (#17)
To give visual distinction for labels I thought to color them, but
having to set and maintain the colors manually seemed cumbersome.
Instead we take the label string as input to create a hex color from
it. Which means the same label will have the same color consistently. A
little cherry on top, the text color is inverted for some aid with
contrast. A long the way some line height and alignment fixes in the
task

* color labels based on their text input
* invert label text color
* extract Label into element
* fix row gap for smaller screen sizes
* align elements to baseline in TaskListItem
* Address review feedack
* Use consistent monospace font and padding for labels